// frontend/src/lib/api/generated/models/discovery.ts
/**
 * Generated by Orval. Do not edit manually.
 */
import type { Assignment } from "./assignment.ts";

export interface Discovery {
  /** A URL to the JSON Schema for this object. */
  readonly $schema?: string;
  devboxes: Assignment[];
  github_user_id: number;
  protocol: number;
  registry_id: string;
  registry_url?: string;
  revision: string;
}